05_MEDIA_LAG/requirements.txt
# Core Dependencies
opencv-python==4.8.0
ffmpeg-python==0.2.0
numpy==1.24.3
Pillow==10.0.0
pydantic==2.4.2

# API and Network
requests==2.31.0
aiohttp==3.8.5
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Media Processing
moviepy==1.0.3
pytube==15.0.0
python-magic==0.4.27

# Social Media
instabot==0.117.0
tweepy==4.14.0
google-api-python-client==2.108.0

# Database and Storage
sqlalchemy==2.0.23
alembic==1.12.1
python-jose==3.3.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
tqdm==4.66.1
colorlog==6.7.0
psutil==5.9.5
tenacity==8.2.3

# Testing and Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.0 